Relocating the console wrapper.
